Revert "Chore/asad/refactor structure"
