--- app/src/main/java/com/alireza/news/app/NewsApplication.kt
@@ -1,7 +1,0 @@
-package com.alireza.news.app
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class NewsApplication : Application()
