--- app/src/main/java/com/alireza/news/core/Constants.kt
@@ -1,3 +1,0 @@
-package com.alireza.news.core
-
-const val TIMEOUT = 30L
