--- app/src/main/java/com/alireza/news/di/AppModule.kt
@@ -1,10 +1,0 @@
-package com.alireza.news.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-}
