--- app/src/main/java/com/alireza/news/di/NetworkModule.kt
@@ -1,74 +1,0 @@
-package com.alireza.news.di
-
-import com.alireza.news.BuildConfig
-import com.alireza.news.BuildConfig.DEBUG
-import com.alireza.news.core.TIMEOUT
-import com.alireza.news.newsList.data.dataSource.remote.api.NewsApi
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun gson(): Gson = GsonBuilder().create()
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(): OkHttpClient = OkHttpClient()
-        .newBuilder()
-        .addInterceptor(
-            HttpLoggingInterceptor().apply {
-                level =
-                    if (DEBUG)
-                        HttpLoggingInterceptor.Level.BODY
-                    else
-                        HttpLoggingInterceptor.Level.NONE
-            }
-        )
-        .addInterceptor(
-            Interceptor { chain ->
-                chain.proceed(
-                    chain.request().newBuilder()
-                        .addHeader("X-Api-Key", BuildConfig.NEWS_API_ACCESS_KEY)
-                        .build()
-                )
-            }
-        )
-        .connectTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .readTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .writeTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .build()
-
-    @Provides
-    @Singleton
-    fun retrofit(
-        okHttpClient: OkHttpClient,
-        gson: Gson
-    ): Retrofit = Retrofit.Builder()
-        .baseUrl(BuildConfig.BASE_URL)
-        .addCallAdapterFactory(CoroutineCallAdapterFactory.invoke())
-        .addConverterFactory(GsonConverterFactory.create(gson))
-        .client(okHttpClient)
-        .build()
-
-
-    @Provides
-    @Singleton
-    fun provideNewsApi(retrofit: Retrofit): NewsApi =
-        retrofit.create(NewsApi::class.java)
-}
