--- app/src/main/java/com/alireza/news/newsList/data/dataSource/local/NewsDao.kt
@@ -1,19 +1,0 @@
-package com.alireza.news.newsList.data.dataSource.local
-
-import androidx.room.Dao
-import androidx.room.Query
-import androidx.room.Upsert
-import com.alireza.news.newsList.data.dataSource.local.entity.NewsArticleEntity
-import kotlinx.coroutines.flow.Flow
-
-
-@Dao
-interface NewsDao {
-
-    @Upsert
-    suspend fun insertNewsArticles(articles: List<NewsArticleEntity>)
-
-    @Query("SELECT * FROM news_articles")
-    fun getNewsArticles(): Flow<List<NewsArticleEntity>>
-
-}
