+++ app/src/main/java/com/alireza/news_app/NewsApplication.kt
@@ -1,0 +1,7 @@
+package com.alireza.news_app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class NewsApplication : Application()
