+++ app/src/main/java/com/alireza/news_app/core/Constants.kt
@@ -1,0 +1,3 @@
+package com.alireza.news_app.core
+
+const val TIMEOUT = 30L
