+++ app/src/main/java/com/alireza/news_app/di/AppModule.kt
@@ -1,0 +1,10 @@
+package com.alireza.news_app.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+}
