+++ app/src/main/java/com/alireza/news_app/feature_news/data/local/NewsDatabase.kt
@@ -1,0 +1,14 @@
+package com.alireza.news_app.feature_news.data.local
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import com.alireza.news_app.feature_news.data.local.entity.NewsArticleEntity
+
+
+@Database(
+    entities = [NewsArticleEntity::class],
+    version = 1,
+)
+abstract class NewsDatabase : RoomDatabase() {
+    abstract val dao: NewsDao
+}
