+++ app/src/main/java/com/alireza/news_app/feature_news/data/local/entity/NewsArticleEntity.kt
@@ -1,0 +1,12 @@
+package com.alireza.news_app.feature_news.data.local.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "news_articles")
+data class NewsArticleEntity(
+    @PrimaryKey(autoGenerate = false)
+    val url: String,
+    val title: String?,
+    val urlToImage: String?
+)
