+++ app/src/main/java/com/alireza/news_app/feature_news/data/remote/NewsApi.kt
@@ -1,0 +1,5 @@
+package com.alireza.news_app.feature_news.data.remote
+
+interface NewsApi {
+
+}
